+++ src/test/java/util/TestEntityFactory.java
@@ -54,7 +54,7 @@ public class TestEntityFactory {
                 .build();
     }
 
-    public static Material createMaterial(String name, QuantityUnit unit, Long price) {
+    public static Material createMaterial(String name, QuantityUnit unit, int price) {
         return Material.builder()
                 .name(name)
                 .baseUnit(unit)
